Integrate Branch class into Diff class
