+++ src/main/java/de/uhd/ifi/se/decision/management/jira/git/model/Diff.java
@@ -1,7 +1,19 @@
 package de.uhd.ifi.se.decision.management.jira.git.model;
 
+import java.net.URLEncoder;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import javax.xml.bind.annotation.XmlElement;
+
+import org.eclipse.jgit.lib.Ref;
+import org.eclipse.jgit.revwalk.RevCommit;
+
+import de.uhd.ifi.se.decision.management.jira.quality.completeness.QualityProblem;
 
 /**
  * Models a list of {@link ChangedFile}s. The scope for the diff might be a
@@ -11,9 +23,27 @@ import java.util.List;
 public class Diff {
 
 	private List<ChangedFile> changedFiles;
+	private List<RevCommit> commits;
+	private Ref ref;
+	private List<DecisionKnowledgeElementInCodeComment> codeElements;
+	private List<DecisionKnowledgeElementInCommitMessage> commitElements;
+	private String repoUri;
 
 	public Diff() {
-		changedFiles = new ArrayList<ChangedFile>();
+		changedFiles = new ArrayList<>();
+		commits = new ArrayList<>();
+	}
+
+	public Diff(List<RevCommit> commits) {
+		this();
+		this.commits = commits;
+	}
+
+	public Diff(Ref ref, List<DecisionKnowledgeElementInCodeComment> codeCommentElements,
+			List<DecisionKnowledgeElementInCommitMessage> commitMessageElements) {
+		this.ref = ref;
+		this.commitElements = commitMessageElements;
+		this.codeElements = codeCommentElements;
 	}
 
 	/**
@@ -50,5 +80,58 @@ public class Diff {
 			elementsFromCode.addAll(codeFile.getRationaleElementsFromCodeComments());
 		}
 		return elementsFromCode;
+	}
+
+	public List<RevCommit> getCommits() {
+		return commits;
+	}
+
+	public void setCommits(List<RevCommit> commits) {
+		this.commits = commits;
+	}
+
+	@XmlElement
+	public String getName() {
+		return ref != null ? ref.getName() : null;
+	}
+
+	@XmlElement
+	public String getId() {
+		return ref != null ? ref.getObjectId().getName() : null;
+	}
+
+	public void setRef(Ref ref) {
+		this.ref = ref;
+	}
+
+	@XmlElement
+	public List<DecisionKnowledgeElementInCodeComment> getCodeElements() {
+		return codeElements;
+	}
+
+	@XmlElement
+	public List<DecisionKnowledgeElementInCommitMessage> getCommitElements() {
+		return commitElements;
+	}
+
+	@XmlElement
+	public String getRepoUri() {
+		return URLEncoder.encode(repoUri, Charset.defaultCharset());
+	}
+
+	public void setRepoUri(String repoUri) {
+		this.repoUri = repoUri;
+	}
+
+	@XmlElement
+	public Set<QualityProblem> getQualityProblems() {
+		return codeElements.stream().flatMap(element -> element.getQualityProblems().stream())
+				.collect(Collectors.toSet());
+	}
+
+	@XmlElement(name = "hash")
+	@Override
+	public int hashCode() {
+		return Objects.hash(getName(), getId());
 	}
 }
